Update Harfbuzz to 7.0.1
This is the 6.4-specific update of Harfbuzz, since Harfbuzz
was currently on an older version in 6.4 compared to other
branches.

This also backports the change to the qtextlayout test that
was done on other branches for the 6.0.0 update.

Fixes: QTBUG-111535
Change-Id: Ie8217f1541a02d3800142b88cf0007be0ed0d1e2
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit 13549fa7c2d27ba588471f97a640f4dbcf86b10c)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>