Increase cache size for QOpenGLTextureCache
Currently images that does not fit in the cache will be destroyed, which
is unfortunate. 256 MB cache is too small for todays standards, so
increasing it to 1 GB. Also adding an environment variable so that it is
changeable if required.

Fixes: QTBUG-111498
Change-Id: I70c65cad6219a59102b16abc50f098aa0b017314
Reviewed-by: Eirik Aavitsland <eirik.aavitsland@qt.io>
(cherry picked from commit 55568b8c626576726c6c9138c8044ea52bc3a355)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>