Doc: fix QOpenGLWidget::defaultFramebufferObject documentation
Fix the link to QSurfaceFormat::StereoBuffers, and clean the text up
a bit. If everything is a \note, then we might just as well not have
notes, so move the statement about when this function is useful out
of a note, and combine the statements elaborating on ownership and
lifetime of framebuffer and contexts into a single note.

Change-Id: I7cb1f9ff78ba760e99d639550130f5c833ad684a
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
(cherry picked from commit 47df0eb5131a5fc0decd049bd3f9eefeb26276df)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>