macOS: Don't set opacity on non-top level windows during NSWindow creation
The opacity is a property of NSWindow.

There's no point in pulling out the opacity from QWindowPrivate,
as QWindow has a public accessor for it. And we don't need to
guard the call with a check for non-1.0 opacity. If we want
logic to avoid redundant calls to the platform APIs, they
should be in the setter.

Change-Id: Ic9b8d1051f30d3b7e09dae14b9f22ca899d05865
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
(cherry picked from commit d7e2c9acf2d84edf8993c81b22a9b4d231aab4ed)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>