Stop propagation and default action on WASM window pointer event
The windows should not let events through to targets below (unless
they are Qt::WindowTransparentForInput, which is handled
independently). Therefore, stopPropagation and preventDefault are
now called regardless of whether the window has handled the event
or not. If the event is not handled explicitly, we still consume the
DOM event so that the windows/elements below don't get focus by
mistake.

Fixes: QTBUG-116557
Change-Id: Idc78ac5076e97992d9f6f13a1fa574d0df4178ac
Pick-to: 6.5.3
Reviewed-by: Morten Johan Sørvig <morten.sorvig@qt.io>