QMetaEnum: fix narrowing warnings related to valueToKey*()
valueToKey*() takes an int, and QMetaEnum internally assumes the
underlying type of an enum is not bigger than int.

Using static_cast as it's easier to search for it in the codebase if
needed.

Found by compiling with -Wshorten-64-to-32.

Drive-by change: remove qdoc \fn command for qt_QMetaEnum_debugOperator,
it's an internal method and defined right under the doc block (the
signature still said "int" but it's qint64).

Change-Id: Ia6abc85173bf94c0a8c56663481d83d3a998f68d
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>